+++ src/components/NavBar.js
@@ -1,0 +1,10 @@
+import Logo from "./Logo.js";
+
+export default function NavBar({ children }) {
+  return (
+    <nav className="nav-bar">
+      <Logo />
+      {children}
+    </nav>
+  );
+}
