+++ src/components/NumResults.js
@@ -1,0 +1,7 @@
+export default function NumResults({ movies }) {
+  return (
+    <p className="num-results">
+      Found <strong>{movies.length}</strong> results
+    </p>
+  );
+}
